refac: Move stuff out of aggregators.py
