+++ source/databricks/calculation-engine/package/steps/aggregation/aggregators.py
@@ -22,138 +22,19 @@ from package.constants import Colname, ResultKeyName
 from . import transformations as T
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import (
-    array_contains,
     col,
-    collect_set,
     lit,
     row_number,
-    sum,
     when,
-    window,
 )
 from pyspark.sql.window import Window
 from typing import Union
 
 
-in_sum = "in_sum"
-out_sum = "out_sum"
-exchange_in_in_grid_area = "ExIn_InMeteringGridArea_Domain_mRID"
-exchange_in_out_grid_area = "ExIn_OutMeteringGridArea_Domain_mRID"
-exchange_out_in_grid_area = "ExOut_InMeteringGridArea_Domain_mRID"
-exchange_out_out_grid_area = "ExOut_OutMeteringGridArea_Domain_mRID"
-
-
-# Function to aggregate net exchange per neighbouring grid areas (step 1)
-def aggregate_net_exchange_per_neighbour_ga(
-    enriched_time_series: DataFrame,
-) -> DataFrame:
-    df = enriched_time_series.where(
-        col(Colname.metering_point_type) == MeteringPointType.exchange.value
-    )
-    exchange_in = (
-        df.groupBy(
-            Colname.in_grid_area,
-            Colname.out_grid_area,
-            Colname.time_window,
-            Colname.aggregated_quality,
-        )
-        .sum(Colname.quantity)
-        .withColumnRenamed(f"sum({Colname.quantity})", in_sum)
-        .withColumnRenamed(Colname.in_grid_area, exchange_in_in_grid_area)
-        .withColumnRenamed(Colname.out_grid_area, exchange_in_out_grid_area)
-    )
-    exchange_out = (
-        df.groupBy(
-            Colname.in_grid_area,
-            Colname.out_grid_area,
-            Colname.time_window,
-        )
-        .sum(Colname.quantity)
-        .withColumnRenamed(f"sum({Colname.quantity})", out_sum)
-        .withColumnRenamed(Colname.in_grid_area, exchange_out_in_grid_area)
-        .withColumnRenamed(Colname.out_grid_area, exchange_out_out_grid_area)
-    )
-
-    exchange = (
-        exchange_in.join(exchange_out, [Colname.time_window], "inner")
-        .filter(
-            exchange_in.ExIn_InMeteringGridArea_Domain_mRID
-            == exchange_out.ExOut_OutMeteringGridArea_Domain_mRID
-        )
-        .filter(
-            exchange_in.ExIn_OutMeteringGridArea_Domain_mRID
-            == exchange_out.ExOut_InMeteringGridArea_Domain_mRID
-        )
-        .select(exchange_in["*"], exchange_out[out_sum])
-        .withColumn(Colname.sum_quantity, col(in_sum) - col(out_sum))
-        .withColumnRenamed(exchange_in_in_grid_area, Colname.in_grid_area)
-        .withColumnRenamed(exchange_in_out_grid_area, Colname.out_grid_area)
-        .withColumnRenamed(Colname.aggregated_quality, Colname.quality)
-        .select(
-            Colname.in_grid_area,
-            Colname.out_grid_area,
-            Colname.time_window,
-            Colname.quality,
-            Colname.sum_quantity,
-            col(Colname.in_grid_area).alias(Colname.grid_area),
-            lit(MeteringPointType.exchange.value).alias(Colname.metering_point_type),
-        )
-    )
-    return T.create_dataframe_from_aggregation_result_schema(exchange)
-
-
-# Function to aggregate net exchange per grid area (step 2)
-def aggregate_net_exchange_per_ga(df: DataFrame) -> DataFrame:
-    exchange_in = df.filter(
-        col(Colname.metering_point_type) == MeteringPointType.exchange.value
-    )
-    exchange_in = (
-        exchange_in.groupBy(
-            Colname.in_grid_area,
-            window(col(Colname.observation_time), "1 hour"),
-            Colname.aggregated_quality,
-        )
-        .sum(Colname.quantity)
-        .withColumnRenamed(f"sum({Colname.quantity})", in_sum)
-        .withColumnRenamed("window", Colname.time_window)
-        .withColumnRenamed(Colname.in_grid_area, Colname.grid_area)
-    )
-    exchange_out = df.filter(
-        col(Colname.metering_point_type) == MeteringPointType.exchange.value
-    )
-    exchange_out = (
-        exchange_out.groupBy(
-            Colname.out_grid_area, window(col(Colname.observation_time), "1 hour")
-        )
-        .sum(Colname.quantity)
-        .withColumnRenamed(f"sum({Colname.quantity})", out_sum)
-        .withColumnRenamed("window", Colname.time_window)
-        .withColumnRenamed(Colname.out_grid_area, Colname.grid_area)
-    )
-    joined = exchange_in.join(
-        exchange_out,
-        (exchange_in[Colname.grid_area] == exchange_out[Colname.grid_area])
-        & (exchange_in[Colname.time_window] == exchange_out[Colname.time_window]),
-        how="outer",
-    ).select(exchange_in["*"], exchange_out[out_sum])
-    result_df = (
-        joined.withColumn(Colname.sum_quantity, joined[in_sum] - joined[out_sum])
-        .withColumnRenamed(Colname.aggregated_quality, Colname.quality)
-        .select(
-            Colname.grid_area,
-            Colname.time_window,
-            Colname.sum_quantity,
-            Colname.quality,
-            lit(MeteringPointType.exchange.value).alias(Colname.metering_point_type),
-        )
-    )
-    return T.create_dataframe_from_aggregation_result_schema(result_df)
-
-
 def aggregate_non_profiled_consumption_ga_brp_es(
     enriched_time_series: DataFrame,
 ) -> DataFrame:
-    return _aggregate(
+    return _aggregate_per_ga_and_brp_and_es(
         enriched_time_series,
         MeteringPointType.consumption,
         SettlementMethod.non_profiled,
@@ -161,7 +42,7 @@ def aggregate_non_profiled_consumption_ga_brp_es(
 
 
 def aggregate_flex_consumption_ga_brp_es(enriched_time_series: DataFrame) -> DataFrame:
-    return _aggregate(
+    return _aggregate_per_ga_and_brp_and_es(
         enriched_time_series,
         MeteringPointType.consumption,
         SettlementMethod.flex,
@@ -169,10 +50,12 @@ def aggregate_flex_consumption_ga_brp_es(enriched_time_series: DataFrame) -> Dat
 
 
 def aggregate_production_ga_brp_es(enriched_time_series: DataFrame) -> DataFrame:
-    return _aggregate(enriched_time_series, MeteringPointType.production, None)
+    return _aggregate_per_ga_and_brp_and_es(
+        enriched_time_series, MeteringPointType.production, None
+    )
 
 
-def _aggregate(
+def _aggregate_per_ga_and_brp_and_es(
     df: DataFrame,
     market_evaluation_point_type: MeteringPointType,
     settlement_method: Union[SettlementMethod, None],
@@ -202,7 +85,7 @@ def _aggregate(
         Colname.energy_supplier_id,
         Colname.time_window,
     ]
-    result = _aggregate_sum_and_set_quality(result, "quarter_quantity", sum_group_by)
+    result = T.aggregate_sum_and_set_quality(result, "quarter_quantity", sum_group_by)
 
     win = Window.partitionBy("GridAreaCode").orderBy(col(Colname.time_window))
 
@@ -261,12 +144,11 @@ def aggregate_flex_consumption_ga_es(results: dict) -> DataFrame:
     )
 
 
-# Function to aggregate sum per grid area and energy supplier (step 12, 13 and 14)
 def _aggregate_per_ga_and_es(
     df: DataFrame, market_evaluation_point_type: MeteringPointType
 ) -> DataFrame:
     group_by = [Colname.grid_area, Colname.energy_supplier_id, Colname.time_window]
-    result = _aggregate_sum_and_set_quality(df, Colname.sum_quantity, group_by)
+    result = T.aggregate_sum_and_set_quality(df, Colname.sum_quantity, group_by)
 
     result = result.select(
         Colname.grid_area,
@@ -300,13 +182,13 @@ def aggregate_flex_consumption_ga_brp(results: dict) -> DataFrame:
     )
 
 
-# Function to aggregate sum per grid area and balance responsible party (step 15, 16 and 17)
+# Function to aggregate sum per grid area and balance responsible party
 def _aggregate_per_ga_and_brp(
     df: DataFrame,
     market_evaluation_point_type: MeteringPointType,
 ) -> DataFrame:
     group_by = [Colname.grid_area, Colname.balance_responsible_id, Colname.time_window]
-    result = _aggregate_sum_and_set_quality(df, Colname.sum_quantity, group_by)
+    result = T.aggregate_sum_and_set_quality(df, Colname.sum_quantity, group_by)
 
     result = result.select(
         Colname.grid_area,
@@ -340,13 +222,13 @@ def aggregate_flex_consumption_ga(results: dict) -> DataFrame:
     )
 
 
-# Function to aggregate sum per grid area (step 18, 19 and 20)
+# Function to aggregate sum per grid area
 def _aggregate_per_ga(
     df: DataFrame,
     market_evaluation_point_type: MeteringPointType,
 ) -> DataFrame:
     group_by = [Colname.grid_area, Colname.time_window]
-    result = _aggregate_sum_and_set_quality(df, Colname.sum_quantity, group_by)
+    result = T.aggregate_sum_and_set_quality(df, Colname.sum_quantity, group_by)
 
     result = result.withColumnRenamed(
         f"sum({Colname.sum_quantity})", Colname.sum_quantity
@@ -359,45 +241,3 @@ def _aggregate_per_ga(
     )
 
     return T.create_dataframe_from_aggregation_result_schema(result)
-
-
-def _aggregate_sum_and_set_quality(
-    result: DataFrame, quantity_col_name: str, group_by: list[str]
-) -> DataFrame:
-    result = result.na.fill(value=0, subset=[quantity_col_name])
-    qualities_col_name = "qualities"
-    result = (
-        result.groupBy(group_by).agg(
-            sum(quantity_col_name).alias(Colname.sum_quantity),
-            collect_set("Quality").alias(qualities_col_name),
-        )
-        # TODO: What about calculated (A06)?
-        .withColumn(
-            "Quality",
-            when(
-                array_contains(
-                    col(qualities_col_name), lit(TimeSeriesQuality.missing.value)
-                )
-                | array_contains(
-                    col(qualities_col_name), lit(TimeSeriesQuality.incomplete.value)
-                ),
-                lit(TimeSeriesQuality.incomplete.value),
-            )
-            .when(
-                array_contains(
-                    col(qualities_col_name),
-                    lit(TimeSeriesQuality.estimated.value),
-                ),
-                lit(TimeSeriesQuality.estimated.value),
-            )
-            .when(
-                array_contains(
-                    col(qualities_col_name),
-                    lit(TimeSeriesQuality.measured.value),
-                ),
-                lit(TimeSeriesQuality.measured.value),
-            ),
-        )
-    )
-
-    return result
