+++ source/databricks/calculation-engine/tests/integration/calculator/steps/aggregation/test_flex_consumption_ga_brp_es.py
@@ -16,7 +16,7 @@ from datetime import datetime
 from package.constants import Colname
 from package.steps.aggregation.aggregators import (
     aggregate_flex_consumption_ga_brp_es,
-    _aggregate,
+    _aggregate_per_ga_and_brp_and_es,
 )
 from package.codelists import (
     MeteringPointType,
@@ -250,7 +250,7 @@ def test_flex_consumption_test_filter_by_domain_is_present(
     time_series_row_factory: Callable[..., DataFrame],
 ) -> None:
     df = time_series_row_factory()
-    aggregated_df = _aggregate(
+    aggregated_df = _aggregate_per_ga_and_brp_and_es(
         df,
         MeteringPointType.consumption,
         SettlementMethod.flex,
@@ -262,7 +262,7 @@ def test_flex_consumption_test_filter_by_domain_is_not_present(
     time_series_row_factory: Callable[..., DataFrame]
 ) -> None:
     df = time_series_row_factory()
-    aggregated_df = _aggregate(
+    aggregated_df = _aggregate_per_ga_and_brp_and_es(
         df,
         MeteringPointType.consumption,
         SettlementMethod.non_profiled,
